requirements.txt
requests==2.23.0
python_dotenv==0.13.0
terminaltables==3.1.0